% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_daynight_times.R
\name{calculate_daynight_times}
\alias{calculate_daynight_times}
\title{Calculate the time (in specified units) one time period overlaps with another}
\usage{
calculate_daynight_times(
  period1_start,
  period1_finish,
  period2_start,
  period2_finish,
  units = "hours"
)
}
\arguments{
\item{period1_start}{The start datetime of the first period}

\item{period1_finish}{The finish datetime of the first period}

\item{period2_start}{The start datetime of the second period}

\item{period2_finish}{The finish datetime of the second period}

\item{units}{Time units to output. Defaults to 'hours'.}
}
\value{
Time in specified units where time periods overlap
}
\description{
Calculate the time (in specified units) one time period overlaps with another
}
