% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_lags.R
\name{create_time_lags}
\alias{create_time_lags}
\title{Create time lagged rows (filled with NA values) using the 'time_column' column
as the starting time.}
\usage{
create_time_lags(
  x,
  n_lag_range = c(1, 14),
  time_lag = days(14),
  lag_amount_units = days(1),
  timezone = "Australia/Perth"
)
}
\arguments{
\item{x}{Your original starting tibble.}

\item{time_lag}{The size of the time lag in a duration (e.g. \code{days()})}

\item{lag_amount_units}{A lubridate object or a numeric in seconds that
determines the lag_amount column's unit.}

\item{timezone}{The timezone for dates}

\item{n_lags}{The number of time lags}
}
\value{
A tibble with additional time-lagged rows
}
\description{
Create time lagged rows (filled with NA values) using the 'time_column' column
as the starting time.
}
