% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gee.R
\name{extract_gee}
\alias{extract_gee}
\title{Extract from google earth engine}
\usage{
extract_gee(
  points,
  collection_name,
  bands,
  scale = 250,
  time_buffer = 16,
  time_summarise_fun = "last",
  debug = FALSE
)
}
\arguments{
\item{time_summarise_fun}{}
}
\description{
Use google earth engine to extract your chosen image collection bands
and then summarise this information for each row in your dataset (\code{points})
}
