% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch data for each row in your dataset using the supplied functions}
\usage{
fetch(
  points,
  ...,
  use_cache = TRUE,
  out_dir = file.path("./output/"),
  out_filename = paste0("output_", format(Sys.time(), "\%Y_\%m_\%d_\%H_\%M_\%S"),
    ".gpkg"),
  cache_dir = file.path(out_dir, "cache/"),
  .time_rep = NA
)
}
\arguments{
\item{points}{A tibble with a \code{sf} "geometry" and a "time_column" (a \code{lubridate} interval),}

\item{...}{Anonymous functions you would like to use on each row of the dataset.}

\item{use_cache}{Whether to cache your progress. Allows you to continue where you left off in case of an error or the process is interrupted.}

\item{out_dir}{A directory to output your result. Is ignored if out_filename = NA.}

\item{out_filename}{The path to output the result. Set to NA to not save the result and only return the result.}

\item{cache_dir}{A directory to output cached progress. Is ignored if use_cache = FALSE.}

\item{.time_rep}{A \code{time_rep} object. Used to repeat data extraction along repeating time intervals relative to the minimum time of each row in the dataset.}
}
\description{
Used to gather all input data necessary for modelling
}
