% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_datetime.R
\name{find_closest_datetime}
\alias{find_closest_datetime}
\title{Find the closest datetime in a vector}
\usage{
find_closest_datetime(dates, x, before = NA)
}
\arguments{
\item{dates}{a vector of dates or datetimes}

\item{x}{a single date or datetime}

\item{before}{Leave as NA (the default) to find the closest or specify a
boolean if the closest datetime should be before x (if TRUE) or after x (if
FALSE)}
}
\value{
A logical vector of length dates.
}
\description{
Finds the closest matching datetime before (before=TRUE), after
(before=FALSE) or either (default or if before=NA)
}
