% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daynight_times.R
\name{get_day_night_hours}
\alias{get_day_night_hours}
\title{get_day_night_hours}
\usage{
get_day_night_hours(start, finish, lat, lon, units = "hours")
}
\arguments{
\item{start}{A POSIXct object representing the start time.}

\item{finish}{A POSIXct object representing the end time.}

\item{lat}{The latitude of the location.}

\item{lon}{The longitude of the location.}

\item{units}{A string representing the units in which to output the time, defaults to 'hours'.}
}
\value{
A list containing the number of day and night hours.
}
\description{
This function calculates the number of day and night hours for a given time range and location.
}
\examples{
\dontrun{
day_night_hours <- get_day_night_hours(start = Sys.time(), finish = Sys.time() + lubridate::hours(24), lat = 51.5074, lon = 0.1278, units = 'minutes')
}
}
