% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_daynight_times.R
\name{get_daynight_times}
\alias{get_daynight_times}
\title{Get day and night time-related information}
\usage{
get_daynight_times(points, savepath = "./output/extracted_day_night_stats.rds")
}
\arguments{
\item{points}{An \code{sf} object that contains a geometry and a column with the
datetime as a \code{lubridate::interval}: \code{time_column}}

\item{savepath}{The path to save your extracted day and night time information
as an rds file.}
}
\value{
the input \code{points} object with day and night time-related information
column-binded.
}
\description{
Calculates the time since sunrise, time since sunset and day and night hours
for your points \code{sf} object
}
