% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daynight_times.R
\name{get_time_since_sunrise}
\alias{get_time_since_sunrise}
\title{get_time_since_sunrise}
\usage{
get_time_since_sunrise(start, lat, lon, units = "hours")
}
\arguments{
\item{start}{A POSIXct object representing the start time.}

\item{lat}{The latitude of the location.}

\item{lon}{The longitude of the location.}

\item{units}{A string representing the units in which to output the time, defaults to 'hours'.}
}
\value{
A numeric representing the time elapsed since sunrise.
}
\description{
This function calculates the time elapsed since sunrise given a specific location and time.
}
\examples{
\dontrun{
time_since_sunrise <- get_time_since_sunrise(
  start = Sys.time(),
  lat = 51.5074,
  lon = 0.1278,
  units = 'minutes'
)
}
}
