% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daynight_times.R
\name{time_in_period}
\alias{time_in_period}
\title{time_in_period}
\usage{
time_in_period(
  period1_start,
  period1_finish,
  period2_start,
  period2_finish,
  units = "hours"
)
}
\arguments{
\item{period1_start}{A POSIXct object representing the start time of the first period.}

\item{period1_finish}{A POSIXct object representing the finish time of the first period.}

\item{period2_start}{A POSIXct object representing the start time of the second period.}

\item{period2_finish}{A POSIXct object representing the finish time of the second period.}

\item{units}{A string representing the units in which to output the time, defaults to 'hours'.}
}
\value{
A numeric representing the time overlap between two periods in specified units.
}
\description{
This function calculates the overlap in time, in specified units, between two
time periods defined by start and end times.
}
\examples{
\dontrun{
overlap_hours <- time_in_period(
  period1_start = as.POSIXct('2023-07-17 06:00:00'),
  period1_finish = as.POSIXct('2023-07-17 18:00:00'),
  period2_start = as.POSIXct('2023-07-17 12:00:00'),
  period2_finish = as.POSIXct('2023-07-17 20:00:00'),
  units = 'hours'
)
}
}
