% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_rep.R
\name{time_rep}
\alias{time_rep}
\title{Create a time_rep object}
\usage{
time_rep(interval, n_before = 0, n_after = 0)
}
\arguments{
\item{interval}{The datetime interval. Should be a lubridate period, e.g.
\code{lubridate::days(14)}.}

\item{n_before}{The number of times this interval should be repeated before the original datetime.}

\item{n_after}{The number of times this interval should be repeated after the original datetime.}
}
\description{
Create a time_rep object
}
