% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_across_times.R
\name{extract_across_times}
\alias{extract_across_times}
\title{Extract across times}
\usage{
extract_across_times(
  points,
  r,
  extract_all_times_at_start = TRUE,
  method = "simple",
  debug = FALSE
)
}
\arguments{
\item{points}{an sf object with the columns "time_column" and "geometry" describing
where and when you want to extract data}

\item{r}{A SpatRaster (from the terra package) with the data you want to extract from}

\item{extract_all_times_at_start}{boolean (for speed optimisation). for best
speed, set to true if there are more different times than there are points.
set to false if there are more points than there are times.}
}
\value{
points with extracted data column binded
}
\description{
Extract across times
}
