% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throw.R
\name{throw}
\alias{throw}
\title{Create a sample of points for extracting data from. In other words, throw.}
\usage{
throw(
  offset = c(-180, -90),
  cellsize = c(10, 10),
  n = c(36, 18),
  time_interval,
  crs = sf::st_crs(4326),
  what = "centers",
  square = TRUE,
  flat_topped = FALSE
)
}
\arguments{
\item{flat_topped}{}
}
\description{
Create a sample of points for extracting data from. In other words, throw.
}
